move cp wrapper into d3d/types.rs where it belongs
